refactor(game-engine): extract handleBlitz + cleanup imports (S27.8.9)
handleBlitz (~250l) deplace dans actions/blitz-action.ts. Le flux
complet est preserve : Foul Appearance, mouvement avec dodge +
Shadowing + Break Tackle, jet d'armure + injury + bounce sur echec,
calcul des assists et pose de pendingBlock.

Cleanup massif des imports orphelins suite a l'extraction :
- rollBlockDiceManyWithRolls (utils/dice)
- calculateOffensiveAssists/DefensiveAssists/BlockDiceCount/Chooser
  (mechanics/blocking ; canBlock + canBlitz restent pour getLegalMoves)
- incrementTeamBlitzCount (game-state ; canTeamBlitz reste)
- checkFoulAppearance (negative-traits)
- applyPickupFailure (failure-helpers ; applyRollFailure reste)
- canDumpOff + getDumpOffReceivers (dump-off ; executeDumpOff reste)
- checkDauntless (dauntless)
- getArmBarBonus (arm-bar)
- logInstablePrevention (negative-traits)

actions.ts 1329 -> 1106 lignes (-223 nettes ; cumul -1600 vs origine
2706 = ~59% reduction). Lint actions.ts + blitz-action.ts : 0
warnings. 4888 tests game-engine passent. Coverage 25.33%.

Reste pour atteindre <= 600 lignes : handleDodge, handleLeap,
handleMove, block subsystem leftovers, getLegalMoves, applyMove.
